% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistfo.R
\name{gistfo}
\alias{gistfo}
\title{Create a private Github gist containing the active RStudio tab and browse to it.
File is named: RStudio_<project>_<filename or file_id>. file_id is a unique id for
untitled files. It does not relate to the untitled number.}
\usage{
gistfo()
}
\value{
nothing.
}
\description{
Create a private Github gist containing the active RStudio tab and browse to it.
File is named: RStudio_<project>_<filename or file_id>. file_id is a unique id for
untitled files. It does not relate to the untitled number.
}
