% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistfo.R
\name{gistfoc}
\alias{gistfoc}
\title{Create a public Github gist containing the active RStudio tab and browse to it,
then open the gist in carbon.now.sh, for easily Tweeting a saucy code pic.
Gist file is named: RStudio_<project>_<filename or file_id>. file_id is a unique id for
untitled files. It does not relate to the untitled number.}
\usage{
gistfoc()
}
\value{
nothing.
}
\description{
Create a public Github gist containing the active RStudio tab and browse to it,
then open the gist in carbon.now.sh, for easily Tweeting a saucy code pic.
Gist file is named: RStudio_<project>_<filename or file_id>. file_id is a unique id for
untitled files. It does not relate to the untitled number.
}
