% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistfo.R
\name{gistfo}
\alias{gistfo}
\title{Create a private Github gist and browse to it.}
\usage{
gistfo()
}
\value{
nothing.
}
\description{
The gist will contain the active text selection or the active RStudio tab.
File is named: \code{RStudio_<project>_<?selection>_<filename or file_id>}.
\code{file_id} is a unique id for untitled files. It does not relate to the
untitled number.
}
