% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistfo.R
\name{gistfo}
\alias{gistfo}
\title{Create a private gist containing the active RStudio tab and browse to it.}
\usage{
gistfo()
}
\value{
nothing.
}
\description{
Create a private gist containing the active RStudio tab and browse to it.
}
